requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-google-genai
gtts
IPython